+++ CHANGELOG.md
@@ -5,11 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Client for ds-storage
 
 ## [1.2.0] - 2022-02-XX
 ### Added
-API service methods renamed (url's).
+- API service methods renamed (url's).
 
 ### Changed
 ID normalisation: